requirements.txt
Flask==3.0.0
tensorflow==2.14.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
prophet==1.1.5
Pillow==10.1.0
opencv-python==4.8.1.78
werkzeug==3.0.1
requests==2.31.0
matplotlib==3.8.2
